backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
sqlalchemy
asyncpg
alembic
pydantic
pydantic-settings
psycopg2
torch
torchvision
pillow
python-multipart
aiofiles